Bug solution multi-vehicles in wamv_gazebo.urdf.xacro (#644)
When trying to launch a second vessel with the name wamv2, the following error message was shown
"[ERROR] [launch]: Caught exception in launch (see debug for traceback): Error Code 20: Msg: Child frame with name[wamv/base_link] specified by joint with name[wamv_external_pivot_joint] not found in model with name[WAM-V]"
The problem was found in line 103 of the wamv_gazebo.urdf.xacro file, in which the namespace was not properly referenced to the namespace but only to the standard topic namespace of "wamv/base_link". As a result without changing the child to "$(arg namespace)/base_link, it is not possible to spawn a second vehicle with a different namespace. Therefor, without the change, it would not be possible to control the vessels independent of each other, as both subscribe to the same topic.

Co-authored-by: Jessica <61019526+j-herman@users.noreply.github.com>
Co-authored-by: M1chaelM <M1chaelM@users.noreply.github.com>
Co-authored-by: Carlos Agüero <caguero@openrobotics.org>